Moved general rules out of views.py
